+++ projects/docs-tools/src/modules/code-examples/code-examples-editor.service.spec.ts
@@ -243,7 +243,7 @@ describe('Code examples editor service', () => {
     expect(() => service.launchEditor(example)).not.toThrow();
   });
 
-  it('should convert "*" versions of SKY UX packages to "^5"', () => {
+  it('should convert "*" versions of SKY UX packages to "^6.0.0-0"', () => {
     codeExample.packageDependencies = {
       '@skyux/foobar': '*',
     };
@@ -251,32 +251,32 @@ describe('Code examples editor service', () => {
     service.launchEditor(codeExample);
 
     expect(stackblitzSpy.calls.mostRecent().args[0].dependencies).toEqual({
-      '@angular/animations': '^12.2.0',
-      '@angular/common': '^12.2.0',
-      '@angular/compiler': '^12.2.0',
-      '@angular/core': '^12.2.0',
-      '@angular/forms': '^12.2.0',
-      '@angular/platform-browser': '^12.2.0',
-      '@angular/platform-browser-dynamic': '^12.2.0',
-      '@angular/router': '^12.2.0',
-      '@skyux/animations': '^5.0.0-0',
-      '@skyux/assets': '^5.0.0-0',
-      '@skyux/config': '^5.0.0-0',
-      '@skyux/core': '^5.0.0-0',
-      '@skyux/errors': '^5.0.0-0',
-      '@skyux/forms': '^5.0.0-0',
-      '@skyux/http': '^5.0.0-0',
-      '@skyux/i18n': '^5.0.0-0',
-      '@skyux/indicators': '^5.0.0-0',
-      '@skyux/layout': '^5.0.0-0',
-      '@skyux/modals': '^5.0.0-0',
-      '@skyux/popovers': '^5.0.0-0',
-      '@skyux/router': '^5.0.0-0',
-      '@skyux/theme': '^5.0.0-0',
-      rxjs: '^6.6.0',
+      '@angular/animations': '^13.0.0',
+      '@angular/common': '^13.0.0',
+      '@angular/compiler': '^13.0.0',
+      '@angular/core': '^13.0.0',
+      '@angular/forms': '^13.0.0',
+      '@angular/platform-browser': '^13.0.0',
+      '@angular/platform-browser-dynamic': '^13.0.0',
+      '@angular/router': '^13.0.0',
+      '@skyux/animations': '^6.0.0-0',
+      '@skyux/assets': '^6.0.0-0',
+      '@skyux/config': '^6.0.0-0',
+      '@skyux/core': '^6.0.0-0',
+      '@skyux/errors': '^6.0.0-0',
+      '@skyux/forms': '^6.0.0-0',
+      '@skyux/http': '^6.0.0-0',
+      '@skyux/i18n': '^6.0.0-0',
+      '@skyux/indicators': '^6.0.0-0',
+      '@skyux/layout': '^6.0.0-0',
+      '@skyux/modals': '^6.0.0-0',
+      '@skyux/popovers': '^6.0.0-0',
+      '@skyux/router': '^6.0.0-0',
+      '@skyux/theme': '^6.0.0-0',
+      rxjs: '^7',
       tslib: '^2.3.0',
       'zone.js': '~0.11.4',
-      '@skyux/foobar': '^5.0.0-0', // <-- Important
+      '@skyux/foobar': '^6.0.0-0', // <-- Important
     });
   });